+++ docs/package.json
@@ -9,15 +9,11 @@
     "check-links": "tsx src/validate-docs-links.ts"
   },
   "devDependencies": {
-    "@octokit/webhooks-definitions": "3.67.3",
     "@types/node": "22.7.8",
-    "@vercel/ncc": "0.38.2",
     "tsx": "4.19.1",
     "typescript": "5.6.3"
   },
   "dependencies": {
-    "@actions/core": "1.11.1",
-    "@actions/github": "5.1.1",
     "github-slugger": "2.0.0",
     "gray-matter": "4.0.3",
     "rehype-raw": "7.0.0",
